// src/models/appModels.ts
export interface DomainModel {
  id: number;
  name: Domain;
}

export enum Domain {
  UXUI_DESIGN = 'UXUI_DESIGN',
  DEVELOPMENT = 'DEVELOPMENT',
  GRAPHIC_DESIGN = 'GRAPHIC_DESIGN',
  WEBMARKETING = 'WEBMARKETING',
  CYBERSECURITY = 'CYBERSECURITY',
  DATA = 'DATA',
  AUDIOVISUAL = 'AUDIOVISUAL',
}

export interface SkillModel {
  id: number;
  name: Skill;
}

export enum Skill {
  USER_RESEARCH = 'USER_RESEARCH',
  WIREFRAMING = 'WIREFRAMING',
  INTERACTIVE_PROTOTYPING = 'INTERACTIVE_PROTOTYPING',
  USER_CENTERED_DESIGN = 'USER_CENTERED_DESIGN',
  INTERACTION_DESIGN = 'INTERACTION_DESIGN',
  WEB_DEVELOPMENT = 'WEB_DEVELOPMENT',
  MOBILE_DEVELOPMENT = 'MOBILE_DEVELOPMENT',
  FRONTEND_DEVELOPMENT = 'FRONTEND_DEVELOPMENT',
  BACKEND_DEVELOPMENT = 'BACKEND_DEVELOPMENT',
  DATABASE_MANAGEMENT = 'DATABASE_MANAGEMENT',

  TYPOGRAPHY = 'TYPOGRAPHY',
  BRAND_IDENTITY_DESIGN = 'BRAND_IDENTITY_DESIGN',
  ILLUSTRATION = 'ILLUSTRATION',
  LAYOUT_DESIGN = 'LAYOUT_DESIGN',
  COLOR_THEORY = 'COLOR_THEORY',

  DIGITAL_STRATEGY = 'DIGITAL_STRATEGY',
  SEO_OPTIMIZATION = 'SEO_OPTIMIZATION',
  SOCIAL_MEDIA_MARKETING = 'SOCIAL_MEDIA_MARKETING',
  CONTENT_MARKETING = 'CONTENT_MARKETING',
  ANALYTICS_TRACKING = 'ANALYTICS_TRACKING',

  NETWORK_SECURITY = 'NETWORK_SECURITY',
  THREAT_ANALYSIS = 'THREAT_ANALYSIS',
  INCIDENT_RESPONSE = 'INCIDENT_RESPONSE',
  VULNERABILITY_ASSESSMENT = 'VULNERABILITY_ASSESSMENT',
  CRYPTOGRAPHY = 'CRYPTOGRAPHY',

  DATA_ANALYSIS = 'DATA_ANALYSIS',
  DATA_VISUALIZATION = 'DATA_VISUALIZATION',
  MACHINE_LEARNING = 'MACHINE_LEARNING',
  DATA_ENGINEERING = 'DATA_ENGINEERING',
  STATISTICAL_MODELING = 'STATISTICAL_MODELING',

  VIDEO_EDITING = 'VIDEO_EDITING',
  MOTION_GRAPHICS = 'MOTION_GRAPHICS',
  SOUND_DESIGN = 'SOUND_DESIGN',
  STORYBOARDING = 'STORYBOARDING',
  CINEMATOGRAPHY = 'CINEMATOGRAPHY',
}

export enum Role {
  OWNER = 'OWNER',
  EDITOR = 'EDITOR',
  READER = 'READER',
}

export type ActionModel = {
  id: number;
  name: string;
  createdAt: Date;
};